requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
click==7.0
et-xmlfile==1.0.1
flask-sqlalchemy==2.3.2
flask==1.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
jdcal==1.4
jinja2==2.10
markupsafe==1.1.0
oauthlib==2.1.0
openpyxl==2.5.12
py-trello==0.14.0
python-dateutil==2.7.5
pytz==2018.7
requests-oauthlib==1.0.0
requests==2.21.0
six==1.12.0
sqlalchemy==1.2.15
urllib3==1.24.1
werkzeug==0.14.1
